+Version 0.4.0
Added support to Google Protobuf. Support for other package sending methods will be dropped on later development.

+++ Assets/Package/Library/NetMessage.cs
@@ -111,7 +111,7 @@ namespace GameWorkstore.NetworkLibrary
             nameof(LobbyReturnToLobby), // 46
         };
 
-        public static string TypeToString(ushort value)
+        public static string TypeToString(int value)
         {
             if (value < 0 || value > Highest)
             {
@@ -128,7 +128,7 @@ namespace GameWorkstore.NetworkLibrary
 
     public class NetMessage
     {
-        public ushort Type;
+        public int Type;
         public NetConnection Conn;
         public NetReader Reader;
         public int ChannelId;
